refactor: replace mane page sections in separate components
