--- src/pages/main/MainPage.module.scss
@@ -1,233 +1,0 @@
-.mainPage {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 40px;
-  max-width: 1280px;
-  padding: 40px;
-  margin: 0 auto;
-
-  .aboutUsContainer {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-    max-width: 1110px;
-    gap: 40px;
-    border-radius: 30px;
-    box-shadow: var(--shadow-card-primary);
-    // padding: 140px 0;
-    // background: url("../../../public/assets/stones-bg.jpg") no-repeat center / cover;
-
-    img {
-      border-radius: 30px;
-      width: 100%;
-      height: 100%;
-      object-fit: cover;
-    }
-
-    .aboutUsContent {
-      position: absolute;
-      display: flex;
-      flex-direction: column;
-      gap: 25px;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
-      z-index: 1;
-    }
-    .aboutUsText {
-      max-width: 450px;
-      font-size: var(--font-size-middle);
-    }
-  }
-
-  .catalogContainer {
-    max-width: 1110px;
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-
-    .catalog {
-      display: grid;
-      grid-template-columns: repeat(2, 1fr) auto;
-      grid-template-rows: repeat(2, 1fr);
-      gap: 15px;
-
-      div {
-        box-shadow: var(--shadow-card-primary);
-        border-radius: 10px;
-        position: relative;
-        overflow: hidden;
-        cursor: pointer;
-
-        &:hover img {
-          filter: brightness(70%);
-          transform: scale(1.05);
-        }
-
-        img {
-          border-radius: 10px;
-          transition: ease-in-out 0.3s;
-        }
-
-        .cardText {
-          position: absolute;
-          left: 0;
-          bottom: 0;
-          width: 100%;
-          padding: 10px;
-          background-color: var(--color-primary-bg-dark);
-          color: var(--color-secondary-text);
-          font-size: var(--font-size-middle);
-          border-radius: 0 0 10px 10px;
-        }
-      }
-
-      div:first-child {
-        grid-column: 1;
-        grid-row: 1 / span 2;
-        width: 540px;
-        max-height: 380px;
-        margin-right: 20px;
-      }
-
-      div:not(:first-child) {
-        max-width: 255px;
-        max-height: 175px;
-
-        img {
-          width: 100%;
-          height: 100%;
-          object-fit: cover;
-        }
-      }
-    }
-  }
-
-  .specialsContainer {
-    // position: relative;
-    width: 100%;
-
-    h2 {
-      max-width: 1110px;
-      margin: 0 auto;
-      margin-bottom: 20px;
-    }
-
-    .specialsWrapper {
-      width: 100%;
-      margin: 0 auto;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      gap: 15px;
-
-      .arrowNav {
-        cursor: pointer;
-
-        &:hover {
-          transform: scale(1.1);
-          filter: invert(40%) sepia(90%) saturate(1000%);
-        }
-      }
-    }
-    .promoWrapper {
-      position: relative;
-      max-width: 1110px;
-      max-height: 472px;
-      box-shadow: var(--shadow-card-primary);
-      border-radius: 30px;
-
-      img {
-        display: block;
-        width: 100%;
-        height: 100%;
-        object-fit: cover;
-        border-radius: 30px;
-      }
-
-      .gradientImg {
-        position: absolute;
-        top: 0;
-        left: 0;
-        z-index: 1;
-      }
-    }
-
-    .specialsContent {
-      position: absolute;
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      width: 70%;
-      height: 70%;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
-      z-index: 2;
-      font-size: var(--font-size-h3);
-      font-weight: var(--font-weight-bold);
-
-      div:nth-child(1) {
-        font-size: var(--font-size-h2);
-      }
-
-      div:nth-child(3) {
-        span:first-child {
-          font-size: var(--font-size-h1);
-          line-height: var(--font-line-height-secondary);
-        }
-      }
-
-      div:nth-child(4) {
-        display: flex;
-        flex-direction: column;
-        span:nth-child(2) {
-          font-size: var(--font-size-h1);
-          color: var(--color-accent-text);
-          -webkit-text-stroke: 2px var(--color-primary-text);
-        }
-      }
-
-      div:nth-child(3) {
-        max-width: 300px;
-      }
-
-      div:nth-child(1),
-      div:nth-child(4) {
-        justify-content: flex-start;
-      }
-
-      div:nth-child(2),
-      div:nth-child(3) {
-        align-self: flex-end;
-      }
-    }
-  }
-
-  .aboutStoreContainer {
-    max-width: 1110px;
-    display: flex;
-    gap: 30px;
-
-    .aboutStoreText {
-      flex: 0 0 50%;
-      display: flex;
-      flex-direction: column;
-      gap: 10px;
-    }
-
-    .aboutStoreImg {
-      flex: 0 0 50%;
-      display: flex;
-      align-items: center;
-
-      img {
-        width: 100%;
-        border-radius: 10px;
-      }
-    }
-  }
-}
