+++ src/pages/main/index.tsx
@@ -1,0 +1,3 @@
+import MainPage from './ui/layoutMainPage/MainPage';
+
+export default MainPage;
