+++ src/pages/main/ui/aboutUsContainer/AboutUsContainer.module.scss
@@ -1,0 +1,33 @@
+.aboutUsContainer {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  max-width: 1110px;
+  gap: 40px;
+  border-radius: 30px;
+  box-shadow: var(--shadow-card-primary);
+
+  img {
+    border-radius: 30px;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+
+  .aboutUsContent {
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    gap: 25px;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    z-index: 1;
+  }
+  .aboutUsText {
+    max-width: 450px;
+    font-size: var(--font-size-middle);
+  }
+}
