+++ src/pages/main/ui/catalogContainer/CatalogContainer.module.scss
@@ -1,0 +1,67 @@
+.catalogContainer {
+  max-width: 1110px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+
+  .catalog {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr) auto;
+    grid-template-rows: repeat(2, 1fr);
+    gap: 15px;
+
+    div {
+      box-shadow: var(--shadow-card-primary);
+      border-radius: 10px;
+      position: relative;
+      overflow: hidden;
+      cursor: pointer;
+
+      &:hover img {
+        filter: brightness(70%);
+        transform: scale(1.05);
+      }
+
+      &:hover .cardText {
+        color: var(--color-accent-text);
+      }
+
+      img {
+        border-radius: 10px;
+        transition: ease-in-out 0.3s;
+      }
+
+      .cardText {
+        position: absolute;
+        left: 0;
+        bottom: 0;
+        width: 100%;
+        padding: 10px;
+        background-color: var(--color-primary-bg-dark);
+        color: var(--color-secondary-text);
+        font-size: var(--font-size-middle);
+        border-radius: 0 0 10px 10px;
+        transition: ease-in-out 0.3s;
+      }
+    }
+
+    div:first-child {
+      grid-column: 1;
+      grid-row: 1 / span 2;
+      width: 540px;
+      max-height: 380px;
+      margin-right: 20px;
+    }
+
+    div:not(:first-child) {
+      max-width: 255px;
+      max-height: 175px;
+
+      img {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+      }
+    }
+  }
+}
