+++ src/pages/main/ui/layoutMainPage/MainPage.module.scss
@@ -1,0 +1,9 @@
+.mainPage {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 40px;
+  max-width: 1280px;
+  padding: 40px;
+  margin: 0 auto;
+}
