+++ src/pages/main/ui/specialsContainer/SpecialsContainer.module.scss
@@ -1,0 +1,100 @@
+.specialsContainer {
+  width: 100%;
+
+  h2 {
+    max-width: 1110px;
+    margin: 0 auto;
+    margin-bottom: 20px;
+  }
+
+  .specialsWrapper {
+    width: 100%;
+    margin: 0 auto;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 15px;
+
+    .arrowNav {
+      cursor: pointer;
+
+      &:hover {
+        transform: scale(1.1);
+        filter: invert(40%) sepia(90%) saturate(1000%);
+      }
+    }
+  }
+
+  .promoWrapper {
+    position: relative;
+    max-width: 1110px;
+    max-height: 472px;
+    box-shadow: var(--shadow-card-primary);
+    border-radius: 30px;
+
+    img {
+      display: block;
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+      border-radius: 30px;
+    }
+
+    .gradientImg {
+      position: absolute;
+      top: 0;
+      left: 0;
+      z-index: 1;
+    }
+  }
+
+  .specialsContent {
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    width: 70%;
+    height: 70%;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    z-index: 2;
+    font-size: var(--font-size-h3);
+    font-weight: var(--font-weight-bold);
+
+    div:nth-child(1) {
+      font-size: var(--font-size-h2);
+    }
+
+    div:nth-child(3) {
+      span:first-child {
+        font-size: var(--font-size-h1);
+        line-height: var(--font-line-height-secondary);
+      }
+    }
+
+    div:nth-child(4) {
+      display: flex;
+      flex-direction: column;
+      span:nth-child(2) {
+        font-size: var(--font-size-h1);
+        color: var(--color-accent-text);
+        -webkit-text-stroke: 2px var(--color-primary-text);
+      }
+    }
+
+    div:nth-child(3) {
+      max-width: 300px;
+    }
+
+    div:nth-child(1),
+    div:nth-child(4) {
+      justify-content: flex-start;
+    }
+
+    div:nth-child(2),
+    div:nth-child(3) {
+      align-self: flex-end;
+    }
+  }
+}
